Moving back Scope to where it was originally (opensearch_internal_data)
