+++ lib/charms/opensearch/v0/opensearch_distro.py
@@ -17,7 +17,6 @@ from typing import Dict, List, Optional, Set, Union
 
 import requests
 import urllib3.exceptions
-from charms.opensearch.v0.constants_charm import Scope
 from charms.opensearch.v0.constants_secrets import ADMIN_PW
 from charms.opensearch.v0.helper_cluster import Node
 from charms.opensearch.v0.helper_conf_setter import YamlConfigSetter
@@ -32,6 +31,7 @@ from charms.opensearch.v0.opensearch_exceptions import (
     OpenSearchHttpError,
     OpenSearchStartTimeoutError,
 )
+from charms.opensearch.v0.opensearch_internal_data import Scope
 
 # The unique Charmhub library identifier, never change it
 LIBID = "7145c219467d43beb9c566ab4a72c454"
