+++ lib/charms/opensearch/v0/opensearch_tls.py
@@ -19,10 +19,10 @@ import re
 import socket
 from typing import Dict, List, Optional, Tuple
 
-from charms.opensearch.v0.constants_charm import Scope
 from charms.opensearch.v0.constants_tls import TLS_RELATION, CertType
 from charms.opensearch.v0.helper_networking import get_host_public_ip
 from charms.opensearch.v0.opensearch_exceptions import OpenSearchError
+from charms.opensearch.v0.opensearch_internal_data import Scope
 from charms.tls_certificates_interface.v1.tls_certificates import (
     CertificateAvailableEvent,
     CertificateExpiringEvent,
